vmspec-boot: Use PCI as the virtio transport
Using PCI is the way of the future as it gives us a way to hotplug
storage controllers and allows the use of vhost for performance.

Note that there seems to be a strange interaction between Ubuntu 16.04
and its included UEFI and kernel, which prevents PCI from working with
that particular image.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>